+++ sailrobot/include/utilities.hpp
@@ -29,28 +29,28 @@ class Utility{
 		 * @{
 		 */
 		static float GPSDist(float lat1, float lon1, float lat2, float lon2);
-		static float GPSDist(glm::vec2 point1, glm::vec2 point2);
+		static float GPSDist(glm::dvec2 point1, glm::dvec2 point2);
 		static float GPSDistFast(float lat1, float lon1, float lat2, float lon2);
-		static float GPSDistFast(glm::vec2 point1, glm::vec2 point2);
+		static float GPSDistFast(glm::dvec2 point1, glm::dvec2 point2);
 		/** @} */
 
 		/** Calculate the initial bearing between two GPS coordinates
 		 * @{
 		 */
 		static float GPSBearing(float lat1, float lon1, float lat2, float lon2);
-		static float GPSBearing(glm::vec2 point1, glm::vec2 point2);
+		static float GPSBearing(glm::dvec2 point1, glm::dvec2 point2);
 		/** @} */
 
 		/** Transforms GPS coordinates to Cartesian coordinates
 		 * @{
 		 */
-		static glm::vec3 GPSToCartesian(float lat, float gpslong);
+		static glm::dvec3 GPSToCartesian(float lat, float gpslong);
 		/** @} */
 
 		/** Transforms GPS coordinates to Cartesian coordinates
 		 * @{
 		 */
-		static glm::vec3 GPSToCartesian(glm::vec2 gpsposition);
+		static glm::dvec3 GPSToCartesian(glm::dvec2 gpsposition);
 		/** @} */
 
 		static float CartesianToGPS(float x, float y);
@@ -58,22 +58,22 @@ class Utility{
 		/** Conversion between Quaternion to Euler and Euler to Quaternion
 		 * @{
 		 */
-		static glm::vec3 QuaternionToEuler(glm::quat q);
-		static glm::vec3 QuaternionToEuler(float x, float y, float z, float w);
-		static glm::vec3 QuaternionToEuler(geometry_msgs::Quaternion q);
-		static glm::quat EulerToQuaternion(glm::vec3 v);
+		static glm::dvec3 QuaternionToEuler(glm::quat q);
+		static glm::dvec3 QuaternionToEuler(float x, float y, float z, float w);
+		static glm::dvec3 QuaternionToEuler(geometry_msgs::Quaternion q);
+		static glm::quat EulerToQuaternion(glm::dvec3 v);
 		static glm::quat EulerToQuaternion(float x, float y, float z);
 		/** @} */
 
 		/// Read a GPS file. The GPS file must have each GPS coordinates (latitude, longitude) separated by an end line.
-		static glm::vec2* ReadGPSCoordinates(std::string filepath, int& size);
-		static glm::vec2* AppendGPSCoordinates(std::string filepath, int& size, glm::vec2* list, int sizeList);
-		static std::vector<glm::vec2> AppendGPSCoordinates(std::string filepath, int& size, std::vector<glm::vec2>* list);
+		static glm::dvec2* ReadGPSCoordinates(std::string filepath, int& size);
+		static glm::dvec2* AppendGPSCoordinates(std::string filepath, int& size, glm::dvec2* list, int sizeList);
+		static std::vector<glm::dvec2> AppendGPSCoordinates(std::string filepath, int& size, std::vector<glm::dvec2>* list);
 
 		static std::map<std::string,std::string> ReadConfig(std::string filepath);
 		static void SaveConfig(std::string filepath);
 
-		static float RelativeToTrueWind(glm::vec2 v, float heading, float windDirection, float windAccx, float windAccy, float* windNorthAcc = NULL);
+		static float RelativeToTrueWind(glm::dvec2 v, float heading, float windDirection, float windAccx, float windAccy, float* windNorthAcc = NULL);
 
 		/** Tacking Strategy
 		 * @{
@@ -85,7 +85,7 @@ class Utility{
 		/** Standard rudder and sail control from Heading
 		 * @{
 		 */
-		static glm::vec2 StandardCommand(glm::vec3 currentHeading, float heading, float windNorth, float max_sail = M_PI/2.0, float max_rudder = M_PI/4.0);
+		static glm::dvec2 StandardCommand(glm::dvec3 currentHeading, float heading, float windNorth, float max_sail = M_PI/2.0, float max_rudder = M_PI/4.0);
 		/** @} */
 
 		std::map<std::string,std::string> config;
